refactor(router): migrate middleware, metrics, and server lifecycle to echo v5
- Register the custom JSONSerializer and drop removed Echo fields
  (HideBanner/HidePort) in newEcho.
- Swap echo-contrib/echoprometheus for github.com/labstack/echo-prometheus and
  otelecho for github.com/labstack/echo-opentelemetry.
- Fix middleware for v5: gzip swaps the embedded ResponseWriter via
  UnwrapResponse; access_logging/request_counter read status/size via
  UnwrapResponse; invoke errors through Echo().HTTPErrorHandler; rate limiter
  Rate is now float64.
- Rework server start/shutdown to use echo.StartConfig with a cancelable
  context since Echo.Shutdown was removed in v5.
- go mod tidy drops echo/v4, echo-contrib, and otelecho.

+++ router/middlewares/rate_limit.go
@@ -14,7 +14,7 @@ import (
 func RateLimit(limit rate.Limit, burst int, expiresIn time.Duration) echo.MiddlewareFunc {
 	return middleware.RateLimiterWithConfig(middleware.RateLimiterConfig{
 		Store: middleware.NewRateLimiterMemoryStoreWithConfig(middleware.RateLimiterMemoryStoreConfig{
-			Rate:      limit,
+			Rate:      float64(limit),
 			Burst:     burst,
 			ExpiresIn: expiresIn,
 		}),
